Moving menus to js/menus directory
